[refactor] NewsListScreen: move to keep ViewModel at NavGraph; create UIEvent data class
