+++ app/src/main/java/uk/ryanwong/skycatnews/newslist/ui/NewsListUIState.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2022. Ryan Wong (hello@ryanwong.co.uk)
  */
 
-package uk.ryanwong.skycatnews.newslist.ui.viewmodel
+package uk.ryanwong.skycatnews.uk.ryanwong.skycatnews.newslist.ui
 
 import uk.ryanwong.skycatnews.app.util.ErrorMessage
 import uk.ryanwong.skycatnews.newslist.domain.model.NewsItem
